New exception added + removed old code
A new exception has been added. This is the OutOfBounds exception. It
is raised when a ship is trying to be placed outside the bounds of the
game board.
Old parameters from the constructor method of the Board class has also
been removed. These parameters were no longer in use and were relics
from when the constructor also generated the game board.
A small amount of documentation has been added in relation to the new
exception and  at the top of the file detailing the purpose of this
program.

Signed-off-by: Computroniks <mnickson@sidingsmedia.com>