Fixed help display + help auto update
The output of the help file has been modified to prevent the printing of
a new line in between every line.
The auto update function for the help file has also been implemented.
A carriage return has been added to the anykey function to allow for
overwriting of the prompt.

Signed-off-by: Matthew Nickson <mnickson@sidingsmedia.com>