Removed timeout + change settings + cleanup
The timeout has been replaced with a check to make sure the board is big
enough. This is because the timeout somtimes failed to create the game
board properly.
The ability to change the apps settings has now been added.
The code has been cleaned up slightly and a few minor bugs in the game
save features have been rectified.

Signed-off-by: Matthew Nickson <mnickson@sidingsmedia.com>